+++ app/components/certificate-item/styles.scss
@@ -1,11 +1,10 @@
 @import "../../styles/values_importer";
 
 td{
+  @include font(12px, 300);
   padding: 15px;
   text-align: left;
   vertical-align:middle;
-  font-weight: 300;
-  font-size: 12px;
   color: $dark-grey;
   border-bottom: solid 1px rgba(255,255,255,0.4);
 }
