+++ app/components/certificate-item/template.hbs
@@ -2,4 +2,4 @@
   <td>{{certificate.name}}</td>
   <td>{{certificate.number}}</td>
   <td>{{issueDateFormatted}}</td>
-  <td>{{issueDateFormatted}}</td>
+  <td>{{expiryDateFormatted}}</td>
