+++ app/components/certificates-table/template.hbs
@@ -1,0 +1,27 @@
+<section>
+  <h2>{{title}}</h2>
+  <div class="tbl-header">
+    <table cellpadding="0" cellspacing="0" border="0">
+      <thead>
+      <tr>
+        <th>Type</th>
+        <th>Name</th>
+        <th>Number</th>
+        <th>Issue date</th>
+        <th>Expiry Date</th>
+      </tr>
+      </thead>
+    </table>
+  </div>
+  <div class="tbl-content">
+    <table cellpadding="0" cellspacing="0" border="0">
+      <tbody>
+      {{#each certificates as |certificate|}}
+        {{certificate-item certificate=certificate session=session}}
+      {{else}}
+        <p class='no-cert'>You have no documents of that type</p>
+      {{/each}}
+      </tbody>
+    </table>
+  </div>
+</section>
