+++ app/styles/app.scss
@@ -4,6 +4,7 @@
 @import 'ember-paper';
 @import 'app/components/certificate-item/styles';
 @import 'app/components/ui/search-input/styles';
+@import 'app/components/certificates-table/styles';
 @import 'app/styles/certificates';
 
 $font-1x: 14px;
@@ -14,6 +15,7 @@ html {
 }
 
 body {
+  height: auto;
   padding-top: 20px;
   background: -webkit-linear-gradient(left, #e2dddd, #b2b2b2);
   background: linear-gradient(to right, #e2dddd, #b2b2b2);
@@ -24,6 +26,22 @@ a {
     text-decoration: none;
     color: $color-snow;
   }
+}
+
+h1 {
+  @include font(30px, 300);
+  color: $dark-grey;
+  text-transform: uppercase;
+  text-align: center;
+  margin-bottom: 15px;
+}
+
+h2 {
+  @include font(20px, 200);
+  color: $dark-grey;
+  text-transform: uppercase;
+  text-align: left;
+  margin-bottom: 5px;
 }
 
 .md-toolbar-tools {
