+++ app/styles/certificates.scss
@@ -4,36 +4,38 @@ section{
   margin: 50px;
 }
 
-h1{
-  font-size: 30px;
-  color: $dark-grey;
-  text-transform: uppercase;
-  font-weight: 300;
-  text-align: center;
-  margin-bottom: 15px;
-}
-
 table{
-  width:100%;
+  width: 100%;
   table-layout: fixed;
 }
 
 .tbl-header{
-  background-color: rgba(255,255,255,0.3);
+  background-color: $color-snow-light;
 }
 
 .tbl-content{
-  height:300px;
-  overflow-x:auto;
+  height: 300px;
+  overflow-x: auto;
   margin-top: 0;
-  border: 1px solid rgba(255,255,255,0.3);
+  border: 1px solid $color-snow-light;
 }
 
 th{
+  @include font(12px, 500);
   padding: 20px 15px;
   text-align: left;
-  font-weight: 500;
-  font-size: 12px;
   color: $dark-grey;
   text-transform: uppercase;
+}
+
+#passports .tbl-content {
+  height: 150px;
+}
+
+#medicalCerts .tbl-content {
+  height: 150px;
+}
+
+#endorsements .tbl-content {
+  height: 150px;
 }
