+++ app/styles/mixins/color_scheme.scss
@@ -2,4 +2,5 @@
 $color-brand-alert: #E80040;
 $color-btn-blue: #29323c;
 $color-snow: rgba(255, 255, 255, 0.87);
+$color-snow-light: rgba(255, 255, 255, 0.3);
 $dark-grey: #222;
