+++ app/templates/certificates/index.hbs
@@ -18,30 +18,8 @@
     {{/toolbar.tools}}
   {{/sticky-container}}
 {{/paper-toolbar}}
-<section>
-  <h1>Certificates</h1>
-  <div class="tbl-header">
-    <table cellpadding="0" cellspacing="0" border="0">
-      <thead>
-      <tr>
-        <th>Type</th>
-        <th>Name</th>
-        <th>Number</th>
-        <th>Issue date</th>
-        <th>Expiry Date</th>
-      </tr>
-      </thead>
-    </table>
-  </div>
-  <div class="tbl-content">
-    <table cellpadding="0" cellspacing="0" border="0">
-      <tbody>
-      {{#each sortedCertificates as |certificate|}}
-          {{certificate-item certificate=certificate session=session}}
-      {{else}}
-        <p>You have no certificates</p>
-      {{/each}}
-      </tbody>
-    </table>
-  </div>
-</section>
+<h1>Certificates</h1>
+{{certificates-table certificates=stcwCertificates title='STCW certificates'}}
+{{certificates-table certificates=passports title="Passports and Seaman's books" id='passports'}}
+{{certificates-table certificates=medicalCertificates title='Medical certificates' id='medicalCerts'}}
+{{certificates-table certificates=endorsements title='Endorsements' id='endorsements'}}
